ReplicatedPG: take and drop read locks when doing backfill
All our interfaces are in place, so now we can actually take and
drop the locks. We do so in a few different places:
1) Take locks in ReplicatedPG::recover_backfill. This is the main
entry into the code path.
1a) Take locks when responding to a pull request. This can only happen
on non-primary nodes, so we don't need to worry about the request blocking.
2) Drop the locks in ReplicatedBackend::build_push_op() when it's done
reading the object and has pushed it all.

Signed-off-by: Greg Farnum <greg@inktank.com>