OSD: maintain the largest epoch received from peers in their Session
We have a problem with fast dispatch: a peer can send us a needed map,
then a subop, and we can queue up the map message in the DispatchQueue
before immediately processing the subop.
osd.0 --> MOSDMap(10,11) --> osd.1
osd.0 --> MOSDSubOp(11)  --> osd.1
      osd.1: DispatchQueue(MOSDMap)
      osd.1: fast_dispatch(MOSDSubOp);
      osd.1: ms_dispatch(MOSDMap);
Whenever we get a message too new for us to process, we must either wait
blindly for a new map to arrive, or preemptively request the new map from
the Monitor. Maintaining a preprocessed largest epoch on the Session lets
us detect the above situation by noting the sent map before we've been
able to incorporate it into the full OSD state. Thus we avoid both bombarding
the monitor with unnecessary traffic and having to make wait time versus
double-work tradeoffs for the common case.

Signed-off-by: Greg Farnum <greg@inktank.com>